python-legacy/services/rating/pyproject.toml
[project]
name = "rating"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "structlog>=24.4",
    "pydantic-settings>=2.0",
    "aio-pika>=9.0",
    "bss-models",
    "bss-clients",
    "bss-clock",
    "bss-events",
    "bss-telemetry",
    "bss-middleware",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "respx>=0.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv.sources]
bss-models = { workspace = true }
bss-clients = { workspace = true }
bss-clock = { workspace = true }
bss-events = { workspace = true }
bss-telemetry = { workspace = true }
bss-middleware = { workspace = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: marks tests requiring real downstream containers",
]
